HELP ME
I keep having this annoying error. I put a void function down and it keeps expecting me to put a semi colon on the end of the bracket. I do that and then it gives another error saying its an incomplete void. I looked it up and that error is because of the semi colon. I tried fixing it by putting () by each void name but then it says the curly bracket to start the void function isnt allowed. NEED HELP